refactor: winnerLotto handles the prize
