+++ src/main/java/lotto/domain/WinnerLotto.java
@@ -3,11 +3,8 @@ package lotto.domain;
 // TODO consider changing to record
 // TODO to package private
 public final class WinnerLotto {
-    // TODO 이거 왜 public임?
-    // 아 지금은 Lotto가 참조하고 있어서 그렇고
-    // 나중에 toPrize 메서드 들여오면 수정하면 됨
-    public Lotto lotto;
-    public int bonusNumber;
+    private Lotto lotto;
+    private int bonusNumber;
 
     public WinnerLotto(Lotto lotto, int bonusNumber) {
         this.lotto = lotto;
@@ -17,18 +14,14 @@ public final class WinnerLotto {
 
     private void validate() {
         LottoRule.NumberRangeRule.DEFAULT.validate(this.bonusNumber);
-        if (this.lotto.isContain(this.bonusNumber)) {
+        if (this.lotto.contains(this.bonusNumber)) {
             throw LottoProblem.DUPLITCATE_NUMBER.exception();
         }
     }
 
-    Prize toPrize(Lotto winner) {
-        // Set<Integer> targetNumbers = Set.copyOf(this.numbers);
-        // int matches = (int)
-        // winner.lotto.numbers.stream().filter(targetNumbers::contains).count();
-        // boolean isBonusMatch = targetNumbers.contains(winner.bonusNumber);
-        // return Prize.of(matches, isBonusMatch);
-        // TODO implement
-        throw new UnsupportedOperationException();
+    Prize toPrize(Lotto guess) {
+        int matches = this.lotto.countMatches(guess);
+        boolean isBonusMatch = guess.contains(bonusNumber);
+        return Prize.of(matches, isBonusMatch);
     }
 }
