+++ src/test/java/lotto/domain/LottoTest.java
@@ -51,42 +51,6 @@ class LottoTest {
                 .hasMessageContaining("NUMBER_OUT_OF_RANGE");
     }
 
-    // TODO WinnerLotto 쪽으로 마이그레이션
-    @Test
-    void 로또_2등_당첨_경우() {
-        Lotto purchase = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        WinnerLotto winner = new WinnerLotto(new Lotto(List.of(1, 2, 3, 4, 5, 45)), 6);
-
-        Prize expected = Prize.MATCH_FIVE_WITH_BONUS;
-        Prize actual = purchase.toPrize(winner);
-
-        assertEquals(expected, actual);
-    }
-
-    // TODO WinnerLotto 쪽으로 마이그레이션
-    @Test
-    void 로또_3개_당첨_경우() {
-        Lotto purchase = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        WinnerLotto winner = new WinnerLotto(new Lotto(List.of(1, 2, 3, 43, 44, 45)), 6);
-
-        Prize expected = Prize.MATCH_THREE;
-        Prize actual = purchase.toPrize(winner);
-
-        assertEquals(expected, actual);
-    }
-
-    // TODO WinnerLotto 쪽으로 마이그레이션
-    @Test
-    void 로또_미당첨_경우() {
-        Lotto purchase = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        WinnerLotto winner = new WinnerLotto(new Lotto(List.of(1, 2, 42, 43, 44, 45)), 6);
-
-        Prize expected = Prize.MATCH_NONE;
-        Prize actual = purchase.toPrize(winner);
-
-        assertEquals(expected, actual);
-    }
-
     @Test
     void testToString() {
         Lotto purchase = new Lotto(List.of(8, 21, 23, 41, 42, 43));
