+++ src/test/java/lotto/domain/WinnerLottoTest.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 
@@ -28,4 +29,41 @@ class WinnerLottoTest {
                 .hasMessageStartingWith("[ERROR]")
                 .hasMessageContaining("DUPLITCATE_NUMBER");
     }
+
+    // TODO WinnerLotto 쪽으로 마이그레이션
+    @Test
+    void 로또_2등_당첨_경우() {
+        Lotto purchase = new Lotto(List.of(1, 2, 3, 4, 5, 6));
+        WinnerLotto winner = new WinnerLotto(new Lotto(List.of(1, 2, 3, 4, 5, 45)), 6);
+
+        Prize expected = Prize.MATCH_FIVE_WITH_BONUS;
+        Prize actual = winner.toPrize(purchase);
+
+        assertEquals(expected, actual);
+    }
+
+    // TODO WinnerLotto 쪽으로 마이그레이션
+    @Test
+    void 로또_3개_당첨_경우() {
+        Lotto purchase = new Lotto(List.of(1, 2, 3, 4, 5, 6));
+        WinnerLotto winner = new WinnerLotto(new Lotto(List.of(1, 2, 3, 43, 44, 45)), 6);
+
+        Prize expected = Prize.MATCH_THREE;
+        Prize actual = winner.toPrize(purchase);
+
+        assertEquals(expected, actual);
+    }
+
+    // TODO WinnerLotto 쪽으로 마이그레이션
+    @Test
+    void 로또_미당첨_경우() {
+        Lotto purchase = new Lotto(List.of(1, 2, 3, 4, 5, 6));
+        WinnerLotto winner = new WinnerLotto(new Lotto(List.of(1, 2, 42, 43, 44, 45)), 6);
+
+        Prize expected = Prize.MATCH_NONE;
+        Prize actual = winner.toPrize(purchase);
+
+        assertEquals(expected, actual);
+    }
+
 }
